+++ tests/main_test.go
@@ -1,0 +1,50 @@
+package tests
+
+import (
+	"fmt"
+	"log"
+	"os"
+	"testing"
+
+	"github.com/jmoiron/sqlx"
+	"github.com/ory/dockertest"
+)
+
+var db *sqlx.DB
+
+func TestMain(m *testing.M) {
+	pool, err := dockertest.NewPool("")
+	if err != nil {
+		log.Fatalf("Could not connect to docker: %s", err)
+	}
+
+	resource, err := pool.Run(
+		"postgres", "12", []string{"POSTGRES_DATABASE=postgres", "POSTGRES_PASSWORD=dev"},
+	)
+	if err != nil {
+		log.Fatalf("Could not start resource: %s", err)
+	}
+
+	if err := pool.Retry(func() error {
+		var err error
+		uri := fmt.Sprintf(
+			"postgresql://postgres:dev@localhost:%s/postgres?sslmode=disable",
+			resource.GetPort("5432/tcp"),
+		)
+		db, err = sqlx.Open("postgres", uri)
+		if err != nil {
+			return err
+		}
+		return db.Ping()
+	}); err != nil {
+		log.Fatalf("Could not connect to docker: %s", err)
+	}
+
+	code := m.Run()
+
+	if err := pool.Purge(resource); err != nil {
+		log.Fatalf("Could not purge resource: %s", err)
+	}
+
+	os.Exit(code)
+}
